--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,4 @@
-bin
+bin/*
+!bin/.gitkeep
 *.o
 .DS_Store