###### Confirm functionality before merge _(same avalanche result)_

Results before:
```
Benchmarking by hashing 50000 keys with a size of 64 bytes

Slash
  408.340ns per hash
  6.380ns per byte
  359 cycles per hash
  5.609 cycles per byte
```

Results after:
```
Benchmarking by hashing 50000 keys with a size of 64 bytes

Slash
  326.620ns per hash
  5.103ns per byte
  97 cycles per hash
  1.516 cycles per byte
```

### Changes:
- includes #4 _(merge #4 first, before #5)_
- moved from `gcc-7` to `gcc` _(as there are no special requirements to the compiler)_
- added `-O3` flag for optimization level 3 _(main speedup improvement)_
- changed `switch` statement `case 1` to `default`, which reduces the comparisons for the hash

### Notes:
- Test case required to validate hash results _(pre-defined set of hashes)_