@oltdaniel 2018-08-06
@kbrsh Updated with upstream and solved logic error